test/pytest: set event loop policy from conftest.py
Having a side-effecting statement like this at the top-level of a random
test file is very strange.

pytest-asyncio gives us the ability to override the `event_loop`
fixture, which not only gives us the ability to return a custom type of
loop, but also essentially gives us the ability to write a co-routine
automatically wrapping each individual async test case.

That's powerful: we can manually create the loop without messing with
global state, but we can also take this as an opportunity to do some of
the same polyfills that systemd_ctypes.async_run() would do (which is
even easier, given the `monkeypatch` fixture).

We also take the chance to ensure that our code for tracking task and
process leaks runs on every async testcase, instead of relying on our
current ad-hoc approach of trying to remember to do it.  This finds two
process leaks (of bash) in the subprocess tests, so fix those up.
